Refactor: extract shared validation utils, DynamicList component, and dedup CSS
- Extract isValidEmail, isValidUrl, hasDisallowedPhoneChars, normalizePhoneForComparison,
  sanitizeOtherLabel, OTHER_LABEL_MAX, and findDuplicates into contactValidation.ts
- Extract DynamicList component and useDragReorder hook into DynamicList.tsx;
  used by GlobalTab (with drag reorder) and AddContactModal (without)
- Replace verbose Map-based useMemo duplicate-detection blocks with findDuplicates()
- Remove NewProjectModal.css (all 30 lines were exact duplicates of Modal.css toggle styles)
- Remove !important from border-radius and opacity rules in global.css and ColumnHeader.css
