--- src/renderer/src/shell/NewProjectModal.css
@@ -1,30 +1,0 @@
-.modal-field-toggle {
-  display: flex;
-  flex-direction: column;
-  gap: 6px;
-}
-
-.modal-toggle-label {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  font-size: 13px;
-  font-weight: 400;
-  color: var(--color-text);
-  cursor: pointer;
-  user-select: none;
-}
-
-.modal-toggle-label input[type='checkbox'] {
-  width: 16px;
-  height: 16px;
-  cursor: pointer;
-}
-
-.modal-toggle-hint {
-  font-size: 12px;
-  color: var(--color-text-muted);
-  margin: 0;
-  line-height: 1.55;
-  padding-left: 24px;
-}
